cli: separate disable command into own module
